+++ cocos/core/pipeline/pass-batched-buffer.ts
@@ -10,6 +10,7 @@ import { Mat4 } from '../math';
 import { SubModel } from '../renderer/scene/submodel';
 import { IRenderObject, UBOLocalBatched } from './define';
 import { Pass } from '../renderer';
+import { BindingLayoutPool, PSOCIPool, PSOCIView } from '../renderer/core/memory-pools';
 
 export class PassBatchedBuffer extends BatchedBuffer {
     private static _buffers = new Map<Pass, BatchedBuffer>();
@@ -27,8 +28,8 @@ export class PassBatchedBuffer extends BatchedBuffer {
         let vbSize = 0;
         let vbIdxSize = 0;
         const vbCount = flatBuffers[0].count;
-        const psoCreateInfo = subModel.psoInfos[passIndx];
-        const bindingLayout = psoCreateInfo.bindingLayout;
+        const psoCI = subModel.psoInfos[passIndx];
+        const bindingLayout = BindingLayoutPool.get(PSOCIPool.get(psoCI, PSOCIView.BINDING_LAYOUT));
         let isBatchExist = false;
         for (let i = 0; i < this.batches.length; ++i) {
             const batch = this.batches[i];
@@ -76,10 +77,10 @@ export class PassBatchedBuffer extends BatchedBuffer {
 
                     // update world matrix
                     Mat4.toArray(batch.uboData.view, ro.model.transform.worldMatrix, UBOLocalBatched.MAT_WORLDS_OFFSET + batch.mergeCount * 16);
-                    if (!batch.mergeCount && batch.psoCreateInfo !== psoCreateInfo) {
+                    if (!batch.mergeCount && batch.psoCI !== psoCI) {
                         bindingLayout.bindBuffer(UBOLocalBatched.BLOCK.binding, batch.ubo);
                         bindingLayout.update();
-                        batch.psoCreateInfo = psoCreateInfo;
+                        batch.psoCI = psoCI;
                     }
 
                     ++batch.mergeCount;
@@ -150,7 +151,7 @@ export class PassBatchedBuffer extends BatchedBuffer {
 
         this.batches.push({
             mergeCount: 1,
-            vbs, vbDatas, vbIdx, vbIdxData, vbCount, ia, ubo, uboData, psoCreateInfo,
+            vbs, vbDatas, vbIdx, vbIdxData, vbCount, ia, ubo, uboData, psoCI,
         });
     }
 }
